Trait zip_extensions::read::ZipArchiveExtensions

pub trait ZipArchiveExtensions {
    fn extract(&mut self, path: &PathBuf) -> ZipResult<()>;
    fn extract_file(
        &mut self,
        file_number: usize,
        destination_file_path: &PathBuf,
        overwrite: bool
    ) -> ZipResult<()>;
    fn extract_file_to_memory(
        &mut self,
        file_number: usize,
        buffer: &mut Vec<u8>
    ) -> ZipResult<()>;
    fn entry_path(&mut self, file_number: usize) -> ZipResult<PathBuf>;
    fn file_number(&mut self, entry_path: &PathBuf) -> Option<usize>;
}

Required Methods

fn extract(&mut self, path: &PathBuf) -> ZipResult<()>

Extracts the current archive to the given directory path.

fn extract_file(
&mut self,
file_number: usize,
destination_file_path: &PathBuf,
overwrite: bool
) -> ZipResult<()>

Extracts an entry in the zip archive to a file.

fn extract_file_to_memory(
&mut self,
file_number: usize,
buffer: &mut Vec<u8>
) -> ZipResult<()>

Extracts an entry in the ZIP archive to the given memory buffer.

fn entry_path(&mut self, file_number: usize) -> ZipResult<PathBuf>

Gets an entry´s path.

fn file_number(&mut self, entry_path: &PathBuf) -> Option<usize>

Finds the index of the specified entry.